refactor(haz): moved evaluator to code-analysis
#1
